third_party/WebKit/LayoutTests/web-animations-api/animations-responsive-motionPosition.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id='element'></div>
<script>

var element = document.getElementById('element');

test(function() {
    var keyframes = [
        {motionPosition: '4%'},
        {motionPosition: '12%'}
    ];

    element.style.width = '100';
    element.style.height = '200';
    var player = element.animate(keyframes, 20);
    player.pause();
    player.currentTime = 15;
    element.style.width = '300';
    element.style.height = '600';
    assert_equals(getComputedStyle(element).motionPosition, '10%');
}, 'Motion position percentages are supported');

test(function() {
    var keyframes = [
        {motionPosition: '8em'},
        {motionPosition: '16em'}
    ];

    element.style.fontSize = '10px';
    var player = element.animate(keyframes, 20);
    player.pause();
    player.currentTime = 5;
    element.style.fontSize = '20px';
    assert_equals(getComputedStyle(element).motionPosition, '200px');
}, 'Motion position lengths respond to style changes');

</script>
